PID制御のある程度の仕様固め

+++ Japan_test-teensy40/include/variable.h
@@ -24,9 +24,9 @@ public:
 Status LCD, white;
 Timer LINE, position;
 
-double val_I = 0;
-double deviation, old_deviation, val_D = 0;
-double operation_A, operation_B, operation_C = 0;
+int val_I = 0;
+int deviation, old_deviation, val_D = 0;
+float operation_A, operation_B, operation_C = 0;
 
 int motor[] = {0, 0, 0, 0};
